requirements.txt
# Биржа и данные
ccxt==4.3.89
pandas==2.2.3
numpy==1.26.4

# Технические индикаторы
ta==0.11.0

# ML
xgboost==2.1.3
scikit-learn==1.5.2
joblib==1.4.2

# NLP / сентимент
transformers==4.46.3
torch==2.5.1
sentencepiece==0.2.0

# База данных
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Telegram-бот
python-telegram-bot==21.7

# Расписание задач
apscheduler==3.10.4

# Парсинг новостей
feedparser==6.0.11
requests==2.32.3

# Конфиг
python-dotenv==1.0.1

# Утилиты
loguru==0.7.2
